requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
decorator==4.4.2
ffmpeg-python==0.2.0
future==0.18.3
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.25.2
Pillow==10.0.0
proglog==0.1.10
pytube==15.0.0
pytube3==9.6.4
requests==2.31.0
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
